native file checking

+++ DistributedFileServer/src/DistributedFileApp/FileManager.java
@@ -5,18 +5,17 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.util.HashMap;
 
-
 public class FileManager {
 
-	//hashmap keeps track of how many clients are reading each file
+	// hashmap keeps track of how many clients are reading each file
 	static HashMap<String, Integer> clientsReading = new HashMap<String, Integer>();
 	static HashMap<String, Integer> fileVersionNumbers = new HashMap<String, Integer>();
-	
+
 	public static boolean fileExistsLocally(String fileName) {
 		String path = System.getProperty("user.home") + "/files/" + fileName;
 		return (new File(path)).exists();
 	}
-	
+
 	public static boolean fileExistsRemotely(String fileName, DistributedFileImpl server) {
 		String fileContents = server.searchOtherServers(fileName);
 		if (fileContents.equalsIgnoreCase("File Not Found")) {
@@ -26,7 +25,7 @@ public class FileManager {
 			return true;
 		}
 	}
-	
+
 	private static void copyToLocal(String fileName, String fileContents) {
 		try {
 			File file = new File(System.getProperty("user.home") + "/files/" + fileName);
@@ -38,58 +37,37 @@ public class FileManager {
 			e.printStackTrace();
 		}
 	}
-	
+
 	public static void addClientReading(String fileName, boolean isNewFile) {
 		if (isNewFile) {
-			clientsReading.put(fileName, 1); //add to hashmap
+			clientsReading.put(fileName, 1); // add to hashmap
+			System.out.println("reader adder");
 		} else {
 			clientsReading.put(fileName, clientsReading.get(fileName) + 1);
+			System.out.println("number reading:" + clientsReading.get(fileName));
+
 		}
 	}
-	
-//	public static String retrieveFromRemoteAndGetFileName(DistributedFileImpl server, String fileName) {
-////		String realFileName = addVersionNumberToFileName(fileName);
-//
-//		String fileString = server.searchOtherServers(fileName); //get file contents from remote server
-//		
-//		createFile(fileString); //create it so client can open
-//		
-//		
-//		return fileName;
-//	}
-	
-//	private static String addVersionNumberToFileName(String inputFileName) {
-//		return inputFileName;//change this when we do version numbers
-//	}
 
 	static void doneReading(String inputFileName) {
-		//decrement number of clients reading in hashmap
+		// decrement number of clients reading in hashmap
 		clientsReading.put(inputFileName, clientsReading.get(inputFileName) - 1);
-		
-		//if no more readers, delete file and remove hashmap entry
-		if (clientsReading.get(inputFileName) == 0) {
+		System.out.println(
+				"number of readers reading " + inputFileName + " is now: " + clientsReading.get(inputFileName));
+
+		// if no more readers, delete file and remove hashmap entry
+		if (clientsReading.get(inputFileName) == 0 && !NativeFileReader.isNativeFile(inputFileName)) {
 			clientsReading.remove(inputFileName);
-			
-			//delete the local file
+			System.out.println("no one is reading " + inputFileName + " so it was deleted.");
+
+			// delete the local file
 			File fileToDelete = new File(inputFileName);
 			fileToDelete.delete();
-		}		
+		}
 	}
 
 	public static HashMap<String, Integer> getClientsReading() {
 		return clientsReading;
 	}
-
-//	private static void createFile(String realFileName) {
-//		File file = new File(realFileName);
-//		try {
-//			file.createNewFile();
-//		} catch (IOException e) {
-//			System.out.println("unable to create file");
-//			e.printStackTrace();
-//		}
-//	}
-	
-
 
 }
